Create criterion modules
