+++ src/otx/algo/detection/yolox.py
@@ -11,7 +11,7 @@ from otx.algo.common.losses import CrossEntropyLoss, L1Loss
 from otx.algo.detection.backbones import CSPDarknet
 from otx.algo.detection.detectors import SingleStageDetector
 from otx.algo.detection.heads import YOLOXHead
-from otx.algo.detection.losses import IoULoss
+from otx.algo.detection.losses import IoULoss, YOLOXCriterion
 from otx.algo.detection.necks import YOLOXPAFPN
 from otx.algo.detection.utils.assigners import SimOTAAssigner
 from otx.algo.utils.support_otx_v1 import OTXv1Helper
@@ -184,14 +184,24 @@ class YOLOXTINY(YOLOX):
             num_classes=num_classes,
             in_channels=96,
             feat_channels=96,
+            train_cfg=train_cfg,
+            test_cfg=test_cfg,
+        )
+        criterion = YOLOXCriterion(
+            num_classes=num_classes,
             loss_cls=CrossEntropyLoss(use_sigmoid=True, reduction="sum", loss_weight=1.0),
             loss_bbox=IoULoss(mode="square", eps=1e-16, reduction="sum", loss_weight=5.0),
             loss_obj=CrossEntropyLoss(use_sigmoid=True, reduction="sum", loss_weight=1.0),
             loss_l1=L1Loss(reduction="sum", loss_weight=1.0),
+        )
+        return SingleStageDetector(
+            backbone=backbone,
+            neck=neck,
+            bbox_head=bbox_head,
+            criterion=criterion,
             train_cfg=train_cfg,
             test_cfg=test_cfg,
         )
-        return SingleStageDetector(backbone, bbox_head, neck=neck, train_cfg=train_cfg, test_cfg=test_cfg)
 
 
 class YOLOXS(YOLOX):
@@ -221,14 +231,24 @@ class YOLOXS(YOLOX):
             num_classes=num_classes,
             in_channels=128,
             feat_channels=128,
+            train_cfg=train_cfg,
+            test_cfg=test_cfg,
+        )
+        criterion = YOLOXCriterion(
+            num_classes=num_classes,
             loss_cls=CrossEntropyLoss(use_sigmoid=True, reduction="sum", loss_weight=1.0),
             loss_bbox=IoULoss(mode="square", eps=1e-16, reduction="sum", loss_weight=5.0),
             loss_obj=CrossEntropyLoss(use_sigmoid=True, reduction="sum", loss_weight=1.0),
             loss_l1=L1Loss(reduction="sum", loss_weight=1.0),
+        )
+        return SingleStageDetector(
+            backbone=backbone,
+            neck=neck,
+            bbox_head=bbox_head,
+            criterion=criterion,
             train_cfg=train_cfg,
             test_cfg=test_cfg,
         )
-        return SingleStageDetector(backbone, bbox_head, neck=neck, train_cfg=train_cfg, test_cfg=test_cfg)
 
 
 class YOLOXL(YOLOX):
@@ -253,14 +273,24 @@ class YOLOXL(YOLOX):
         bbox_head = YOLOXHead(
             num_classes=num_classes,
             in_channels=256,
+            train_cfg=train_cfg,
+            test_cfg=test_cfg,
+        )
+        criterion = YOLOXCriterion(
+            num_classes=num_classes,
             loss_cls=CrossEntropyLoss(use_sigmoid=True, reduction="sum", loss_weight=1.0),
             loss_bbox=IoULoss(mode="square", eps=1e-16, reduction="sum", loss_weight=5.0),
             loss_obj=CrossEntropyLoss(use_sigmoid=True, reduction="sum", loss_weight=1.0),
             loss_l1=L1Loss(reduction="sum", loss_weight=1.0),
+        )
+        return SingleStageDetector(
+            backbone=backbone,
+            neck=neck,
+            bbox_head=bbox_head,
+            criterion=criterion,
             train_cfg=train_cfg,
             test_cfg=test_cfg,
         )
-        return SingleStageDetector(backbone, bbox_head, neck=neck, train_cfg=train_cfg, test_cfg=test_cfg)
 
 
 class YOLOXX(YOLOX):
@@ -290,11 +320,21 @@ class YOLOXX(YOLOX):
             num_classes=num_classes,
             in_channels=320,
             feat_channels=320,
+            train_cfg=train_cfg,
+            test_cfg=test_cfg,
+        )
+        criterion = YOLOXCriterion(
+            num_classes=num_classes,
             loss_cls=CrossEntropyLoss(use_sigmoid=True, reduction="sum", loss_weight=1.0),
             loss_bbox=IoULoss(mode="square", eps=1e-16, reduction="sum", loss_weight=5.0),
             loss_obj=CrossEntropyLoss(use_sigmoid=True, reduction="sum", loss_weight=1.0),
             loss_l1=L1Loss(reduction="sum", loss_weight=1.0),
+        )
+        return SingleStageDetector(
+            backbone=backbone,
+            neck=neck,
+            bbox_head=bbox_head,
+            criterion=criterion,
             train_cfg=train_cfg,
             test_cfg=test_cfg,
         )
-        return SingleStageDetector(backbone, bbox_head, neck=neck, train_cfg=train_cfg, test_cfg=test_cfg)
