+++ src/otx/algo/instance_segmentation/losses/__init__.py
@@ -5,5 +5,6 @@
 
 from .accuracy import accuracy
 from .dice_loss import DiceLoss
+from .rtmdet_inst_loss import RTMDetInstCriterion
 
-__all__ = ["accuracy", "DiceLoss"]
+__all__ = ["accuracy", "DiceLoss", "RTMDetInstCriterion"]
